[Lucas Yonghee Lee] renewal/176: fix #FG-WEBADMIN-API-RENEWAL/176
bug fix

+++ src/main/java/net/fashiongo/webadmin/data/model/statistics/HotSearch.java
@@ -8,8 +8,8 @@ public class HotSearch {
 	private String SearchQuery;
 
 	@JsonProperty("Count")
-	private Long Count;
+	private Integer Count;
 
 	@JsonProperty("rankno")
-	private Integer rankno;
+	private Long rankno;
 }
